created router templates
