+++ server/index.js
@@ -1,3 +1,4 @@
-var router = module.exports = require('express').Router();
+// var router = module.exports = require('express').Router();
+// // router = require('express').Router();
 
-router.use('/test_endpoint', require('./test_endpoint'))
+// router.use('/orgs/create', require('./routes/orgs'))
