+++ server/routes/teams.js
@@ -1,0 +1,19 @@
+const server = require('../server.js');
+const { Router } = require('express');
+const router = Router();
+
+/* ------------- Begin Model Functions ------------- */
+
+
+
+/* ------------- End Model Functions ------------- */
+
+
+/* ------------- Begin Controller Functions ------------- */
+
+
+
+/* ------------- End Controller Functions ------------- */
+
+
+module.exports = router;
